refactor: rename dvsportal_new to dvsportal, remove old dvsportal provider
