--- src/pycityvisitorparking/provider/dvsportal_new/CHANGELOG.md
@@ -1,6 +1,0 @@
-# Changelog
-
-## Unreleased
-
-- Added the refactored `dvsportal_new` provider package with split session, profile,
-  transport, and mapping modules.
