--- src/pycityvisitorparking/provider/dvsportal_new/__init__.py
@@ -1,5 +1,0 @@
-"""Refactored DVS Portal provider package."""
-
-from .api import Provider
-
-__all__ = ["Provider"]
